+++ simworld.h
@@ -377,8 +377,8 @@ private:
 	// Data for maintaining industry density even
 	// after industries close
 	// @author: jamespetts
-	double industry_density_proportion;
-	double actual_industry_density;
+	uint32 industry_density_proportion;
+	uint32 actual_industry_density;
 
 	// what game objectives
 	scenario_t *scenario;
@@ -927,11 +927,11 @@ public:
 	inline bool ist_markiert(const grund_t* gr) const { return marker.ist_markiert(gr); }
 
 	// Getter/setter methods for maintaining the industry density
-	inline double get_target_industry_density() const { return finance_history_month[0][WORLD_CITICENS] * industry_density_proportion; }
-	inline double get_actual_industry_density() const { return actual_industry_density; }
+	inline uint32 get_target_industry_density() const { return finance_history_month[0][WORLD_CITICENS] * industry_density_proportion; }
+	inline uint32 get_actual_industry_density() const { return actual_industry_density; }
 	
-	inline void decrease_actual_industry_density(double value) { actual_industry_density -= value; }
-	inline void increase_actual_industry_density(double value) { actual_industry_density += value; }
+	inline void decrease_actual_industry_density(uint32 value) { actual_industry_density -= value; }
+	inline void increase_actual_industry_density(uint32 value) { actual_industry_density += value; }
 
 	 /**
 	 * Initialize map.
